# generated/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Java httpbin apis"
  version: "1.2.0"
  title: "Java httpbin apis"
  termsOfService: "https://github.com/jibijose/httpbin"
  contact:
    name: "Jibi Jose"
    url: "https://httpbin.org"
    email: "jibijose@yahoo.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "jibici.eastus.azurecontainer.io:8080"
schemes:
- "http"
paths:
...
  /delay/random/{unit}/range/{minTime}/{maxTime}:
    get:
      operationId: "delayRandomRange"
      parameters:
      - name: "unit"
        in: "path"
        required: true
        type: "string"
      - name: "minTime"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - name: "maxTime"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        default:
          description: "successful operation"
  /delay/random/{unit}/{time}:
    get:
      operationId: "delayRandom"
      parameters:
      - name: "unit"
        in: "path"
        required: true
        type: "string"
      - name: "time"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        default:
          description: "successful operation"
  /delay/{unit}/{time}:
    get:
      operationId: "delayConstant"
      parameters:
      - name: "unit"
        in: "path"
        required: true
        type: "string"
      - name: "time"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        default:
          description: "successful operation"
...
  /download/{size}:
    get:
      operationId: "downloadSize"
      produces:
      - "text/plain"
      parameters:
      - name: "size"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "string"
              format: "byte"
...
  /file/other/txt/{size}:
    get:
      operationId: "otherTxtSize"
      produces:
      - "text/plain"
      parameters:
      - name: "size"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "string"
              format: "byte"
...
  /file/other/{fileType}:
    get:
      operationId: "otherFileType"
      produces:
      - "text/csv"
      - "text/html"
      - "text/plain"
      - "application/zip"
      parameters:
      - name: "fileType"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "string"
              format: "byte"
...
  /memory/{bytes}/{time}:
    get:
      operationId: "memoryHold"
      parameters:
      - name: "bytes"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - name: "time"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        default:
          description: "successful operation"
  /request/headers:
    get:
      operationId: "headers"
      produces:
      - "application/json"
      parameters:
      - name: "headers"
        in: "header"
        required: true
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            additionalProperties:
              type: "string"
  /status/random/{codes}:
    get:
      operationId: "statusRandomCodes"
      parameters:
      - name: "codes"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "successful operation"
  /status/{code}:
    get:
      operationId: "statusCode"
      parameters:
      - name: "code"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        default:
          description: "successful operation"
...
  /upload:
    post:
      operationId: "upload"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "file"
        in: "query"
        required: true
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/UploadInfoModel"
definitions:
...
  UploadInfoModel:
    type: "object"
    properties:
      uploadedBytesLength:
        type: "string"
      contentType:
        type: "string"
      name:
        type: "string"
      originalFileName:
        type: "string"
      size:
        type: "string"
      status:
        type: "string"